Ignore error replies in a MULTI/EXEC
The only concern is if the MULTI/EXEC block has been executed. The only
error that can occur after the call to MULTI is for an unknown command
and can therefore safely be ignored (this would mean the client is used
in a way it is not meant to be used). The multi bulk reply returned by
EXEC may contain errors when commands are called against a wrong type of
key (e.g. a list push against a string). However, the EXEC still was
successful, which is what is important.

A side-effect of ignoring error replies inside a MULTI/EXEC block is
that the entire block can be pipelined.