Refactor blocking pop tests to be race-free
These tests included bits to explicitly test the blocking nature of the
blocking pop commands. This was done by blocking the main thread and
pushing from another one, and was prone to race conditiions.

Because it is enough to verify that these commands put something
sensible on the wire, the tests now use a mock server to verify that the
timeout argument is correctly passed.