Fix error triggering unrelated error (synchrony)
When `@req.succeed` was called with an error reply that was never
rescued, this error would bubble back to `@req.succeed`, and be caught
by an unrelated rescue block. This rescue block would try to resume the
now dead fiber with the error it caught, resulting in a `FiberError`.

This change unwraps the `@req.succeed` and `@req.fail` calls from blocks
where errors are rescued, such that any error that they trigger will
properly bubble up the call stack (back to EventMachine).

Fixes #248.