A fix for an issue the prevented STOREing SORTs that contained multiple GETs.
When using multiple GETs, Redis#sort expects the return to be an array.
However, combined with a STORE clause, the SORT will return an integer
representing the number of keys stored in the STORE list. The fix checks
if the reply from the Redis client is an integer, and if it is, it does
not attempt to process it as an array.