Fix off by 2 error in EXEC reply handling
The EXEC reply should have the same number of replies as there are commands in
the MULTI. This line of code expected the MULTI and EXEC commands themselves to
be part of the futures array, but they are not.

Note that this only surfaces for Redis versions < 2.6.5, because that is when
the EXECABORT error reply was introduced. Also see: antirez/redis@41f0f92.