Fix #_write_to_socket in connection
By placing the NBIO_READ_EXCEPTIONS above the NBIO_WRITE_EXCEPTIONS, an
Errno::EWOULDBLOCK or Errno::EAGAIN would trigger a wait for the socket
to be readable. This is causing timeout errors in our testing.

When calling #write_nonblock, the appropriate action is to wait for the
socket to be writable, not readable. Ensuring that we handle the write
exceptions before the read exceptions accomplishes this.