Verify certs by default under TLS
`SSLContext#set_params` is the step in SSL connection establishment that
merges user-provided SSL parameters with "saner defaults".

Previously `SSLContext#set_params` was only being called when
`ssl_params` was provided to the SSLConnection, so the SSLContexet was
left unconfigured. This leads to the surprising default `verify_mode` of
`nil`, resulting in SSL connections verifying hostnames but never
verifying certificate trust.