Implement close method
The connection_pool gem is implementing  automatic reaping of idle connections. In order to do this generically, clients need to provide a `close` method that the pool can call on reaped connections in order to close the network connection cleanly.

mysql2 and dalli already provide `close`. This PR adds this convention to redis-rb.