Expose Redis#queue_key
Prior to this commit, the Redis class maintains a queue
mapping Thread.current.object_id to the current queue of
not-yet-committed commands for the particular thread
accessing the Redis client. This works for the common
case where a Redis client is shared amongst multiple
threads, but with Ruby 3's Fiber Scheduler and recent
strides in the Async gem, we'll want a way to choose a
different concurrency primitive (such as a Fiber-backed
Async task) to map queues to. Previously we exposed
`def synchronize` so that the concurrency scheme for
checking out the Redis client could be overridden in
subclasses, but internally the implicit queue key was
hardwired to Thread.current.object_id. This commit takes
a similar approach be exposing `def cache_key` so that it
too can be overriden by Redis subclasses.